Agent/requirements.txt
asgiref==3.8.1
Django==4.0.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
requests==2.31.0
psutil==5.9.8
jsbeautifier==1.15.1